Move tests and helpers to tests package
